Update README.md
The Modern Periodic Table project is a simple project built without the use of graphics. It is developed using the C programming language for the purpose of storing name, symbol, atomic number, and some important properties as well as to display them as per requirement of the user.
Its key features are that it displays information about the D block elements such as name, symbol, electronic configuration etc.
In the project, if you want to display element information is to be displayed of D block of Modern Periodic Table, you have to enter 1 in the main menu and input information in given format. These information are stored in file created on the hard disk of computer by program itself.
•	Exploration of element Information: Another main function of this project is to explore or to display the stored information. You can search an element by using its atomic number.
•	User can close the periodic table if he wants
•	User can see all elements of D block in periodic table